DATAREDIS-425 - Refactoring Part V (Custom Conversions).
Enable custom conversions for more influence on conversions.

Still needs some polishing aka tests, factoryBeans, documentation and the such but gives a glimpse on what’s the purpose of it.

Important: we also need to check for potential index structures on custom converted objects, which has not been implemented so far.